change JKS file copying plugin
We have used maven-dependency-plugin to extract product distribution and copy JKS files.
But with this approach, if we ran  WUM build, then it's getting failed. The reason is
there could be a situation where cert update in the WUM pack. As a solution, we have to
extract distribution in the distribution/target folder and copy JKS files to tests modules.
So the fix is to use maven-antrun-plugin to overcome the problem.